chore: split up git session across files
